crowdstrike-falcon-azure-sentinel-data-connector-durable/requirements.txt
azure-functions
azure-functions-durable
boto3
requests
asyncio
aiofiles